spring-boot-weather-forecast-client-api/src/main/java/com/weather/forecast/api/service/ProductService.java
package com.weather.forecast.api.service;


import com.weather.forecast.api.model.Product;
import com.weather.forecast.api.repository.ProductRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductService {

    private final ProductRepository productRepository;

    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    public Product save(Product product){
        return productRepository.save(product);
    }

    public List<Product> findAll(){
        return productRepository.findAll();
    }

}